extract hourly table
